pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "secure-email"
version = "0.1.0"
description = "Secure dual-domain email MVP aligned with the PDF implementation plan."
requires-python = ">=3.12"
dependencies = [
  "argon2-cffi>=23.1.0",
  "fastapi>=0.115.0",
  "httpx>=0.28.0",
  "pydantic>=2.9.0",
  "PyYAML>=6.0.2",
  "uvicorn>=0.30.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
]

[project.scripts]
secure-mail-server = "server.main:main"
secure-mail-cli = "client.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.setuptools]
packages = ["client", "common", "server"]
